feat: v[i] on a packed array through a typed view (#2041); Format-tier stdlib docs (#1523); an LP64 pointer-type bug and a fmt_gate leftover

+++ std/collections/aether_collections.h
@@ -211,8 +211,8 @@ void floatarr_free(FloatArray* arr);
 // LongArray — fixed-size packed long-long buffer with O(1) random access.
 //
 // The 64-bit twin of IntArray. Same shape, same bounds-check policy.
-// Aether's `long` lowers to C `long long`, so the element type is
-// `long long` end-to-end. Use this for 64-bit-keyed lookup, packed-word
+// Aether's `long` lowers to C `int64_t`, so the element type is
+// `int64_t` end-to-end. Use this for 64-bit-keyed lookup, packed-word
 // tables (hash state words, crypto round buffers), and any other
 // workload that wants a flat packed-long buffer without going through
 // std.list's void*-boxed-per-entry overhead.
@@ -225,7 +225,7 @@ void floatarr_free(FloatArray* arr);
 LongArray* longarr_new_raw(int size);
 
 // Allocate and fill with `init` at every index.
-LongArray* longarr_new_filled_raw(int size, long long init);
+LongArray* longarr_new_filled_raw(int size, int64_t init);
 
 // Number of elements. Returns -1 if `arr` is NULL — distinguishable
 // from a legal empty array (size 0).
@@ -234,16 +234,16 @@ int longarr_size(LongArray* arr);
 // Read the value at `i`. Returns 0 if `arr` is NULL or `i` is
 // out-of-range. Aether wrapper `longarr.get` turns these into
 // Go-style `(value, err)` returns.
-long long longarr_get_raw(LongArray* arr, int i);
+int64_t longarr_get_raw(LongArray* arr, int i);
 
 // Write `value` at `i`. No-op if `arr` is NULL or `i` is out-of-range.
-void longarr_set_raw(LongArray* arr, int i, long long value);
+void longarr_set_raw(LongArray* arr, int i, int64_t value);
 
 // Hot-path skip-the-bounds-check variants. Caller is responsible for
 // keeping the index in [0, size); OOB is undefined behaviour.
 
 // Fill every element with `value`.
-void longarr_fill(LongArray* arr, long long value);
+void longarr_fill(LongArray* arr, int64_t value);
 
 // Release the backing buffer and the struct. Idempotent on NULL.
 void longarr_free(LongArray* arr);
